# backend/src/main/java/com/bancoias/backend/dto/ErrorResponse.java
package com.bancoias.backend.dto;

public record ErrorResponse(
        ErrorDetalle error
) {
    public record ErrorDetalle(
            String tipo,
            String codigo,
            String mensaje
    ) {
    }
}